// src/modules/social-media/types/social-platform.enum.ts
/**
 * Supported social media platforms
 */
export enum SocialPlatform {
  TWITTER = 'TWITTER',
  FACEBOOK = 'FACEBOOK',
  INSTAGRAM = 'INSTAGRAM',
  TIKTOK = 'TIKTOK',
}

/**
 * Status of a social media post
 */
export enum PostStatus {
  DRAFT = 'DRAFT',
  SCHEDULED = 'SCHEDULED',
  POSTED = 'POSTED',
  FAILED = 'FAILED',
  DELETED = 'DELETED',
}